[fixed] constant values in idl translation [fixed] some warnings
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@1645 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/packets/scene/UpdateTransformMessage.h
@@ -86,7 +86,7 @@ public:
 		insertByte(0); // unknown
 
 		// add direction
-		insertByte(scno->getDirectionAngle());
+		insertByte((byte) scno->getDirectionAngle());
 	}
 
 	static void parse(Packet* pack, SceneObject* scno) {
